dev_requirements.txt
cython
pysparkling
psutil

pre-commit

mock
freezegun
nose==1.3.6
nose-cov==1.6
nose-xunitmp==0.3.2
nose-parameterized==0.5.0
pylint~=1.5.6
